requirements.txt
Markdown==3.2.1
pelican==4.2.0
jupyter==1.0.0
ipython==7.13.0
nbconvert==5.6.1
beautifulsoup4==4.8.2
ghp-import==0.5.5
matplotlib==3.2.1
PyYAML==5.3.1
